Try rearranging some files into a better structure
