+++ packages/server/src/helpers/simpleDigiCredsError.ts
