+++ packages/server/src/index.ts
@@ -1,8 +1,6 @@
-export {
-  type DCAPIRequestOID4VP,
-  type DCAPIRequestOptions,
-  generateRequestOptions,
-} from './generateRequestOptions.ts';
+export { generateRequestOptions } from './generateRequestOptions.ts';
+
+export type { DCAPIRequestOID4VP, DCAPIRequestOptions } from './dcapi.ts';
 
 export type {
   OID4VPClaimQuery,
@@ -11,6 +9,6 @@ export type {
   OID4VPCredentialQueryMdoc,
   OID4VPSupportedMdocClaimName,
   PathPointer,
-} from './oid4vp.ts';
+} from './protocols/oid4vp.ts';
 
-export type { DocumentTypeMDLv1, Identifier, NamespaceMDLv1 } from './mdoc.ts';
+export type { DocumentTypeMDLv1, Identifier, NamespaceMDLv1 } from './formats/mdoc.ts';
